Playwright/requirements.txt
# FastAPI + Uvicorn
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Playwright pour le scraping
playwright>=1.40.0

# Google Sheets API
gspread>=5.12.0
google-auth>=2.23.0

# Utilitaires
python-dateutil>=2.8.2
